Do type checking / unwrapping the right way.
GetAlignedPoniterFromInternalField() can crash if the internal field is not a pointer, so we can't use this trick of checking if the field contains a pointer to the type's type_info.

The correct thing to do is use FunctionTemplate::HasInstance() to type-check the value. To do that, we of course have to have a reference to the FunctionTemplate already. To make this lookup fast, I used template-static-singletons.